+++ cycletls/tests/integration/connection_reuse_test.go
@@ -84,7 +84,7 @@ func TestConnectionReuse(t *testing.T) {
 	}
 
 	// Make multiple requests using the same client instance to test connection reuse
-	client := cycletls.Init(false) // Don't use worker pool to focus on connection reuse
+	client := cycletls.Init(cycletls.WithRawBytes())
 	defer client.Close()           // Ensure resources are cleaned up
 
 	// Make first request
@@ -218,7 +218,7 @@ func TestConnectionReuseDisabled(t *testing.T) {
 	}
 
 	// Make multiple requests using the same client instance to test connection reuse disabled
-	client := cycletls.Init(false) // Don't use worker pool to focus on connection reuse
+	client := cycletls.Init(cycletls.WithRawBytes())
 	defer client.Close()           // Ensure resources are cleaned up
 
 	// Make first request